requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.86.0
h11==0.14.0
httptools==0.5.0
idna==3.4
pip==21.3.1
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
setuptools==60.2.0
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.4.0
uvicorn==0.19.0
watchfiles==0.18.0
websockets==10.4
wheel==0.37.1
oracledb==1.1.1
sqlalchemy==1.4.43
python-decouple==3.6
cx_Oracle==8.3.0
psycopg2==2.9.5